% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C_to_K}
\alias{C_to_K}
\title{Celsius conversion
Convert degrees Celsius temperatures to degrees Kelvin}
\usage{
C_to_K(C_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Kelvin
}
\description{
Celsius conversion
Convert degrees Celsius temperatures to degrees Kelvin
}
\examples{
temp1 <- C_to_K(25);
temp2 <- C_to_K( c(-4, 10, 20) );
}
