% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F_to_K}
\alias{F_to_K}
\title{Farenheit conversion
Convert degrees Farenheit temperatures to Kelvin}
\usage{
F_to_K(F_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in Kelvin
}
\description{
Farenheit conversion
Convert degrees Farenheit temperatures to Kelvin
}
\examples{
temp1 <- F_to_K(40);
temp2 <- F_to_K( c(40, 53, 13) );
}
