% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin conversion
Convert Kelvin temperatures to degrees Fahrenheit}
\usage{
K_to_F(K_temp)
}
\arguments{
\item{K_temp}{The temperature in Kelvin}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Kelvin conversion
Convert Kelvin temperatures to degrees Fahrenheit
}
\examples{
temp1 <- K_to_F(10);
temp2 <- K_to_F( c(50, 233, 113) );
}
